Rename feature detection macros.
It is possible that macros with OPENSSL_ prefix will collide with
OpenSSL's usage. The NEVERBLEED_ prefix will not for sure.

Also, descriptive names will prevent from reusing them in inappropriate
ways, like OPENSSL_1_1_API has been used for two different purpose: 1)
to check if RSA_METHOD is opaque and 2) to check if the new unified EC
APIs are available.

Signed-off-by: Masanori Ogino <masanori.ogino@gmail.com>